Fixed compile warning on 64-bit Mac OS X

+++ test/testplatform.c
@@ -24,23 +24,27 @@ TestTypes(SDL_bool verbose)
 
     if (badsize(sizeof(Uint8), 1)) {
         if (verbose)
-            printf("sizeof(Uint8) != 1, instead = %u\n", sizeof(Uint8));
+            printf("sizeof(Uint8) != 1, instead = %u\n",
+                   (unsigned int)sizeof(Uint8));
         ++error;
     }
     if (badsize(sizeof(Uint16), 2)) {
         if (verbose)
-            printf("sizeof(Uint16) != 2, instead = %u\n", sizeof(Uint16));
+            printf("sizeof(Uint16) != 2, instead = %u\n",
+                   (unsigned int)sizeof(Uint16));
         ++error;
     }
     if (badsize(sizeof(Uint32), 4)) {
         if (verbose)
-            printf("sizeof(Uint32) != 4, instead = %u\n", sizeof(Uint32));
+            printf("sizeof(Uint32) != 4, instead = %u\n",
+                   (unsigned int)sizeof(Uint32));
         ++error;
     }
 #ifdef SDL_HAS_64BIT_TYPE
     if (badsize(sizeof(Uint64), 8)) {
         if (verbose)
-            printf("sizeof(Uint64) != 8, instead = %u\n", sizeof(Uint64));
+            printf("sizeof(Uint64) != 8, instead = %u\n",
+                   (unsigned int)sizeof(Uint64));
         ++error;
     }
 #else
